Now works with background-size: cover
Found a way to calculate how much the browser is scaling the image, which means we can now scale the the SVG and fully support background-size: cover. Also cleaned up some comments, and now instead of reading attributes every time we resize, we load them on initialization and store them in the config for better performance.